Clean up code and documentation
