+++ samples/connectivity/telecom/src/main/java/com/example/platform/connectivity/telecom/model/TelecomCallAction.kt
@@ -21,6 +21,13 @@ import android.os.Parcelable
 import android.telecom.DisconnectCause
 import kotlinx.parcelize.Parcelize
 
+/**
+ * Simple interface to represent related call actions to communicate with the registered call scope
+ * in the [TelecomCallRepository.registerCall]
+ *
+ * Note: we are using [Parcelize] to make the actions parcelable so they can be directly used in the
+ * call notification.
+ */
 sealed interface TelecomCallAction : Parcelable {
     @Parcelize
     object Answer : TelecomCallAction
@@ -35,11 +42,11 @@ sealed interface TelecomCallAction : Parcelable {
     object Activate : TelecomCallAction
 
     @Parcelize
-    data class Mute(val isMute: Boolean) : TelecomCallAction
+    data class ToggleMute(val isMute: Boolean) : TelecomCallAction
 
     @Parcelize
     data class SwitchAudioType(val type: Int) : TelecomCallAction
 
     @Parcelize
-    data class TransferCall(val id: ParcelUuid) : TelecomCallAction
+    data class TransferCall(val endpointId: ParcelUuid) : TelecomCallAction
 }
